+++ src/Server.cpp
@@ -1,5 +1,4 @@
-
-#include "Server.hpp"
+#include "./include/Server.hpp"
 
 Server::Server(){
 	this->port = "9034";
